[Game-1] Added a main menu that you start from and return to
Added a simple scene with a button and a script for loading scenes.
Added a return button in the game.
Updated the InitializeMaps function to check if its already been initialized as now it gets loaded every time you come in and out of the game from the main menu.